+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 ### Improved
-- `process.env.NODE_ENV` and `process.env.NODE_DEBUG` are now supported using a webpack `DefinePlugin`
+- `process.env.NODE_ENV` is now set to `development` using a webpack `DefinePlugin`
 
 ## 12.0.0 - 2021-09-06